backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# PDF Processing
PyMuPDF==1.24.9

# Embeddings & Vector Store
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4

# LLM - Google Gemini
google-generativeai==0.8.1

# Utilities
pydantic==2.9.1
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
httpx==0.27.2
